guest will go to login

+++ routes/web.php
@@ -30,29 +30,29 @@ Route::redirect('/home', 'videos')->name('home');
 Route::get('/admin-users', function () {
     return view('users-admin');
 });
-Route::get('/users-admin', [ProfileController::class, 'index'])->name('users-admin');
-Route::put('/users/{user}', [ProfileController::class, 'update'])->name('users.update');
+Route::get('/users-admin', [ProfileController::class, 'index'])->name('users-admin')->middleware('auth');;
+Route::put('/users/{user}', [ProfileController::class, 'update'])->name('users.update')->middleware('auth');;
 
 
-Route::get('/user/{user}', [UserController::class, 'show'])->name('user.profile');
-Route::get('/user/{user}/subscriptions', [UserController::class, 'subscriptions'])->name('user.subscriptions');
-Route::get('/user/{user}/subscribers', [UserController::class, 'subscribers'])->name('user.subscribers');
+Route::get('/user/{user}', [UserController::class, 'show'])->name('user.profile')->middleware('auth');;
+Route::get('/user/{user}/subscriptions', [UserController::class, 'subscriptions'])->name('user.subscriptions')->middleware('auth');;
+Route::get('/user/{user}/subscribers', [UserController::class, 'subscribers'])->name('user.subscribers')->middleware('auth');;
 
-Route::get('/videos', [VideoController::class, 'index'])->name('videos.index');
-Route::get('/videos-user', [VideoController::class, 'userVideos'])->name('videos-user');
-Route::get('/videos/create', [VideoController::class, 'create'])->name('videos.create');
-Route::post('/videos', [VideoController::class, 'store'])->name('videos.store');
-Route::get('/videos/{video}', [VideoController::class, 'show'])->name('videos.show');
-Route::get('/videos/{video}/edit', [VideoController::class, 'edit'])->name('videos.edit');
-Route::put('/videos/{video}', [VideoController::class, 'update'])->name('videos.update');
-Route::delete('/videos/{video}', [VideoController::class, 'destroy'])->name('videos.destroy');
+Route::get('/videos', [VideoController::class, 'index'])->name('videos.index')->middleware('auth');;
+Route::get('/videos-user', [VideoController::class, 'userVideos'])->name('videos-user')->middleware('auth');;
+Route::get('/videos/create', [VideoController::class, 'create'])->name('videos.create')->middleware('auth');;
+Route::post('/videos', [VideoController::class, 'store'])->name('videos.store')->middleware('auth');;
+Route::get('/videos/{video}', [VideoController::class, 'show'])->name('videos.show')->middleware('auth');;
+Route::get('/videos/{video}/edit', [VideoController::class, 'edit'])->name('videos.edit')->middleware('auth');;
+Route::put('/videos/{video}', [VideoController::class, 'update'])->name('videos.update')->middleware('auth');;
+Route::delete('/videos/{video}', [VideoController::class, 'destroy'])->name('videos.destroy')->middleware('auth');;
 
 
 Route::post('/subscribe/{user}', [SubscriptionController::class, 'subscribe'])->name('subscribe')->middleware('auth');
 // Route::post('/unsubscribe/{user}', [SubscriptionController::class, 'unsubscribe'])->name('unsubscribe')->middleware('auth');
-Route::delete('/unsubscribe/{user}', [SubscriptionController::class, 'unsubscribe'])->name('unsubscribe');
+Route::delete('/unsubscribe/{user}', [SubscriptionController::class, 'unsubscribe'])->name('unsubscribe')->middleware('auth');;
 
 
-Route::get('/profile/{user}', [ProfileController::class, 'show'])->name('profile.show');
+Route::get('/profile/{user}', [ProfileController::class, 'show'])->name('profile.show')->middleware('auth');;
